add crop_model to util
